@@ -34,28 +34,26 @@
     "storybook": "storybook dev --quiet -p 7001"
   },
   "devDependencies": {
-    "@eslint/js": "9.26.0",
+    "@eslint/js": "9.28.0",
     "@formatjs/intl-datetimeformat": "6.18.0",
     "@formatjs/intl-numberformat": "8.15.4",
-    "@navikt/aksel-stylelint": "7.21.1",
+    "@navikt/aksel-stylelint": "7.22.0",
     "@navikt/ft-frontend-storybook-utils": "workspace:^",
-    "@storybook/addon-a11y": "8.6.12",
-    "@storybook/addon-actions": "8.6.12",
-    "@storybook/addon-essentials": "8.6.12",
-    "@storybook/addon-links": "8.6.12",
-    "@storybook/builder-vite": "8.6.12",
-    "@storybook/cli": "8.6.12",
-    "@storybook/react": "8.6.12",
-    "@storybook/react-vite": "8.6.12",
-    "@storybook/theming": "8.6.12",
+    "@storybook/addon-a11y": "9.0.4",
+    "@storybook/addon-docs": "9.0.4",
+    "@storybook/addon-links": "9.0.4",
+    "@storybook/builder-vite": "9.0.4",
+    "@storybook/cli": "9.0.4",
+    "@storybook/react": "9.0.4",
+    "@storybook/react-vite": "9.0.4",
     "@testing-library/jest-dom": "6.6.3",
-    "@types/react": "19.1.4",
+    "@types/react": "19.1.6",
     "@types/react-dom": "19.1.5",
     "@types/react-responsive": "9.0.0",
-    "@vitejs/plugin-react": "4.4.1",
+    "@vitejs/plugin-react": "4.5.0",
     "@vitest/coverage-c8": "0.33.0",
     "cross-env": "7.0.3",
-    "eslint": "9.26.0",
+    "eslint": "9.28.0",
     "eslint-config-prettier": "10.1.5",
     "eslint-plugin-import": "2.31.0",
     "eslint-plugin-jsx-a11y": "6.10.2",
@@ -64,7 +62,7 @@
     "eslint-plugin-simple-import-sort": "12.1.1",
     "eslint-plugin-vitest": "0.5.4",
     "gh-pages": "6.3.0",
-    "globals": "16.1.0",
+    "globals": "16.2.0",
     "husky": "9.1.7",
     "jsdom": "26.1.0",
     "lerna": "8.2.2",
@@ -72,15 +70,15 @@
     "postcss-less": "6.0.0",
     "prettier": "3.5.3",
     "react-test-renderer": "19.1.0",
-    "storybook": "8.6.12",
-    "stylelint": "16.19.1",
+    "storybook": "9.0.4",
+    "stylelint": "16.20.0",
     "stylelint-order": "7.0.0",
     "typescript": "5.8.3",
-    "typescript-eslint": "8.32.1",
+    "typescript-eslint": "8.33.0",
     "vite": "6.3.5",
     "vite-plugin-circular-dependency": "0.5.0",
-    "vite-plugin-dts": "4.5.3",
-    "vitest": "3.1.3"
+    "vite-plugin-dts": "4.5.4",
+    "vitest": "3.1.4"
   },
   "resolutions": {
     "@types/react": "19.0.3",
