@@ -1,7 +1,7 @@
 {
   "name": "@navikt/ft-fakta-beregning",
   "description": "Faktaavklaringer rundt aktiviteter og andeler som skal bidra til beregningsgrunnlaget",
-  "version": "9.0.5",
+  "version": "9.0.6",
   "license": "MIT",
   "type": "module",
   "types": "./dist/index.d.ts",
@@ -22,9 +22,9 @@
     "clean": "rm -rf ./dist ./node_modules"
   },
   "dependencies": {
-    "@navikt/aksel-icons": "7.21.1",
-    "@navikt/ds-css": "7.21.1",
-    "@navikt/ds-react": "7.21.1",
+    "@navikt/aksel-icons": "7.22.0",
+    "@navikt/ds-css": "7.22.0",
+    "@navikt/ds-react": "7.22.0",
     "@navikt/ft-form-hooks": "workspace:^",
     "@navikt/ft-form-validators": "workspace:^",
     "@navikt/ft-ui-komponenter": "workspace:^",
@@ -34,7 +34,7 @@
     "dayjs-plugin-utc": "0.1.2",
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "react-hook-form": "7.56.3",
+    "react-hook-form": "7.57.0",
     "react-intl": "7.1.11"
   },
   "peerDependencies": {
@@ -63,11 +63,11 @@
     "@testing-library/react": "16.3.0",
     "@testing-library/user-event": "14.6.1",
     "cross-env": "7.0.3",
-    "eslint": "9.26.0",
-    "stylelint": "16.19.1",
+    "eslint": "9.28.0",
+    "stylelint": "16.20.0",
     "typescript": "5.8.3",
     "vite": "6.3.5",
-    "vitest": "3.1.3"
+    "vitest": "3.1.4"
   },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com"
