@@ -1,7 +1,7 @@
 {
   "name": "@navikt/ft-form-hooks",
   "description": "Overbygg over react-hook-form. Tilbyr form-komponenter som radioknapp og textarea",
-  "version": "6.3.8",
+  "version": "6.3.9",
   "license": "MIT",
   "type": "module",
   "types": "./dist/index.d.ts",
@@ -22,18 +22,18 @@
     "clean": "rm -rf ./dist ./node_modules"
   },
   "dependencies": {
-    "@navikt/aksel-icons": "7.21.1",
-    "@navikt/ds-css": "7.21.1",
-    "@navikt/ds-react": "7.21.1",
-    "@navikt/ft-form-validators": "^3.2.2",
-    "@navikt/ft-ui-komponenter": "^4.4.0",
-    "@navikt/ft-utils": "^3.6.0",
+    "@navikt/aksel-icons": "7.22.0",
+    "@navikt/ds-css": "7.22.0",
+    "@navikt/ds-react": "7.22.0",
+    "@navikt/ft-form-validators": "^3.2.3",
+    "@navikt/ft-ui-komponenter": "^4.4.1",
+    "@navikt/ft-utils": "^3.6.1",
     "classnames": "2.5.1",
     "dayjs": "1.11.13",
     "lodash.throttle": "4.1.1",
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "react-hook-form": "7.56.3"
+    "react-hook-form": "7.57.0"
   },
   "peerDependencies": {
     "@navikt/aksel-icons": "7.x",
@@ -56,11 +56,11 @@
     "@testing-library/react": "16.3.0",
     "@testing-library/user-event": "14.6.1",
     "cross-env": "7.0.3",
-    "eslint": "9.26.0",
-    "stylelint": "16.19.1",
+    "eslint": "9.28.0",
+    "stylelint": "16.20.0",
     "typescript": "5.8.3",
     "vite": "6.3.5",
-    "vitest": "3.1.3"
+    "vitest": "3.1.4"
   },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com"
