@@ -1,7 +1,7 @@
 {
   "name": "@navikt/ft-form-validators",
   "description": "Valideringsregler for bruk i forms",
-  "version": "3.2.2",
+  "version": "3.2.3",
   "license": "MIT",
   "type": "module",
   "types": "./dist/index.d.ts",
@@ -21,7 +21,7 @@
     "clean": "rm -rf ./dist ./node_modules"
   },
   "dependencies": {
-    "@navikt/ft-utils": "^3.6.0",
+    "@navikt/ft-utils": "^3.6.1",
     "dayjs": "1.11.13"
   },
   "peerDependencies": {
@@ -30,10 +30,10 @@
   },
   "devDependencies": {
     "cross-env": "7.0.3",
-    "eslint": "9.26.0",
+    "eslint": "9.28.0",
     "typescript": "5.8.3",
     "vite": "6.3.5",
-    "vitest": "3.1.3"
+    "vitest": "3.1.4"
   },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com"
