@@ -1,6 +1,6 @@
 {
   "name": "@navikt/ft-hooks",
-  "version": "2.1.10",
+  "version": "2.1.11",
   "license": "MIT",
   "type": "module",
   "types": "./dist/index.d.ts",
@@ -23,10 +23,10 @@
     "react": "19.1.0"
   },
   "devDependencies": {
-    "eslint": "9.26.0",
+    "eslint": "9.28.0",
     "typescript": "5.8.3",
     "vite": "6.3.5",
-    "vitest": "3.1.3"
+    "vitest": "3.1.4"
   },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com"
