@@ -1,6 +1,6 @@
 {
   "name": "@navikt/ft-frontend-storybook-utils",
-  "version": "3.2.23",
+  "version": "3.2.24",
   "license": "MIT",
   "type": "module",
   "private": true,
@@ -13,12 +13,12 @@
   "dependencies": {
     "@navikt/ft-form-hooks": "workspace:^",
     "react": "19.1.0",
-    "react-hook-form": "7.56.3",
+    "react-hook-form": "7.57.0",
     "react-intl": "7.1.11"
   },
   "devDependencies": {
-    "@storybook/types": "8.6.12",
-    "eslint": "9.26.0",
+    "@storybook/types": "8.6.14",
+    "eslint": "9.28.0",
     "typescript": "5.8.3"
   }
 }
