@@ -1,6 +1,6 @@
 {
   "name": "@navikt/ft-types",
-  "version": "3.3.1",
+  "version": "3.3.2",
   "license": "MIT",
   "type": "module",
   "private": true,
@@ -22,11 +22,11 @@
   },
   "devDependencies": {
     "cross-env": "7.0.3",
-    "eslint": "9.26.0",
-    "stylelint": "16.19.1",
+    "eslint": "9.28.0",
+    "stylelint": "16.20.0",
     "typescript": "5.8.3",
     "vite": "6.3.5",
-    "vitest": "3.1.3"
+    "vitest": "3.1.4"
   },
   "exports": {
     ".": {
