@@ -3,7 +3,7 @@ import type { StorybookConfig } from '@storybook/react-vite';
 import { lagStoriesEntriesForPakker } from './lagStoriesEntriesForPakker';
 
 const config: StorybookConfig = {
-  addons: ['@storybook/addon-links', '@storybook/addon-essentials', '@storybook/addon-a11y'],
+  addons: ['@storybook/addon-links', '@storybook/addon-a11y', '@storybook/addon-docs'],
   framework: '@storybook/react-vite',
   viteFinal: async c => ({
     ...c,
