@@ -1,5 +1,5 @@
-import { action } from '@storybook/addon-actions';
 import { Meta, StoryObj } from '@storybook/react';
+import { action } from 'storybook/actions';
 
 import { alleKodeverk } from '@navikt/ft-frontend-storybook-utils';
 import { AktivitetStatus, Inntektskategori, OpptjeningAktivitetType, PeriodeAarsak } from '@navikt/ft-kodeverk';
