@@ -2,8 +2,8 @@ import React from 'react';
 import { useForm, UseFormProps } from 'react-hook-form';
 
 import { Button, VStack } from '@navikt/ds-react';
-import { action } from '@storybook/addon-actions';
 import type { Meta } from '@storybook/react';
+import { action } from 'storybook/actions';
 
 import { Form } from './../src/Form';
 
