@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
-import { action } from '@storybook/addon-actions';
 import { Meta, StoryObj } from '@storybook/react';
+import { action } from 'storybook/actions';
 
 import { Step } from './Step';
 import { StepType } from './StepType';
